+++ src/datasources/asset/data-sources/asset-summary/AssetSummaryDataSource.ts
@@ -5,49 +5,52 @@ import { AssetSummaryResponse } from 'datasources/asset/types/AssetSummaryQuery.
 import { AssetDataSourceBase } from '../AssetDataSourceBase';
 import { AssetDataSourceOptions, AssetQuery, AssetQueryType } from '../../types/types';
 import { assetSummaryFields } from '../../constants/AssetSummaryQuery.constants';
+import { catchError, map, Observable, throwError } from 'rxjs';
+
 export class AssetSummaryDataSource extends AssetDataSourceBase {
-    constructor(
-        readonly instanceSettings: DataSourceInstanceSettings<AssetDataSourceOptions>,
-        readonly backendSrv: BackendSrv = getBackendSrv(),
-        readonly templateSrv: TemplateSrv = getTemplateSrv()
-    ) {
-        super(instanceSettings, backendSrv, templateSrv);
-    }
+  constructor(
+    readonly instanceSettings: DataSourceInstanceSettings<AssetDataSourceOptions>,
+    readonly backendSrv: BackendSrv = getBackendSrv(),
+    readonly templateSrv: TemplateSrv = getTemplateSrv()
+  ) {
+    super(instanceSettings, backendSrv, templateSrv);
+  }
 
-    baseUrl = this.instanceSettings.url + '/niapm/v1';
+  baseUrl = this.instanceSettings.url + '/niapm/v1';
 
-    defaultQuery = {
-        type: AssetQueryType.AssetSummary,
+  defaultQuery = {
+    type: AssetQueryType.AssetSummary,
+  };
+
+  runQuery(query: AssetQuery, options: DataQueryRequest): Observable<DataFrameDTO> {
+    return this.getAssetSummary().pipe(
+      map((assetSummary: AssetSummaryResponse) => {
+        return this.processSummaryQuery(query, assetSummary);
+      })
+    );
+  }
+
+  shouldRunQuery(query: AssetQuery): boolean {
+    return !query.hide;
+  }
+
+  processSummaryQuery(query: AssetQuery, assets: AssetSummaryResponse) {
+    return {
+      refId: query.refId,
+      fields: [
+        { name: assetSummaryFields.TOTAL, values: [assets.total] },
+        { name: assetSummaryFields.ACTIVE, values: [assets.active] },
+        { name: assetSummaryFields.NOT_ACTIVE, values: [assets.notActive] },
+        { name: assetSummaryFields.APPROACHING_DUE_DATE, values: [assets.approachingRecommendedDueDate] },
+        { name: assetSummaryFields.PAST_DUE_DATE, values: [assets.pastRecommendedDueDate] }
+      ]
     };
+  }
 
-    async runQuery(query: AssetQuery, options: DataQueryRequest): Promise<DataFrameDTO> {
-        return this.processSummaryQuery(query as AssetQuery);
-    }
-
-    shouldRunQuery(query: AssetQuery): boolean {
-        return !query.hide;
-    }
-
-    async processSummaryQuery(query: AssetQuery) {
-        const assets: AssetSummaryResponse = await this.getAssetSummary();
-
-        return {
-            refId: query.refId,
-            fields: [
-                { name: assetSummaryFields.TOTAL, values: [assets.total] },
-                { name: assetSummaryFields.ACTIVE, values: [assets.active] },
-                { name: assetSummaryFields.NOT_ACTIVE, values: [assets.notActive] },
-                { name: assetSummaryFields.APPROACHING_DUE_DATE, values: [assets.approachingRecommendedDueDate] },
-                { name: assetSummaryFields.PAST_DUE_DATE, values: [assets.pastRecommendedDueDate] }
-            ]
-        };
-    }
-
-    async getAssetSummary(): Promise<AssetSummaryResponse> {
-        try {
-            return await this.get<AssetSummaryResponse>(this.baseUrl + '/asset-summary');
-        } catch (error) {
-            throw new Error(`An error occurred while getting asset summary: ${error}`);
-        }
-    }
+  getAssetSummary(): Observable<AssetSummaryResponse> {
+    return this.get$<AssetSummaryResponse>(this.baseUrl + '/asset-summary').pipe(
+      catchError(error => {
+        return throwError(() => new Error(`An error occurred while getting asset summary: ${error}`));
+      }));
+  }
 }